prompt: Fix exitstatus glitch workaround
Fix it by assuming exit status symbol as double-width using `%{seq%2G%}`
The cost of this "official" workaround is we have to add space if the
symbol is not double-width char.

This is a better workaround than treating it as single-width char
since we'll have to add `\x08` to reset the cursor by one-width, or with
previous hacks which needs two conditional exit status just to fix it.

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>